requirements.txt
asgiref==3.6.0
Django==4.1.5
django-bootstrap5==22.2        
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-crispy-forms==1.14.0    
django-js-asset==2.0.0
django-taggit==3.1.0
django-taggit-serializer==0.1.7
Markdown==3.4.1
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0